Implement similarity compression with overlap
This gets better results than the old similarity compression
algorithm that's in the 6502 assembly code: it allows the similar
sequence to overlap the start position. As long as we decode it
from low index to high index, this still works fine for the decoder
since it will have written the data it's going to copy before
it needs to be read. For this reason, I adjusted the 6502 decoder
to count up on the Y register, which costs an extra 2 cycles per
character decoded.